This commit was manufactured by cvs2svn to create tag 'Ext-5-2-3-pre3'.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/tags/Ext-5-2-3-pre3@14805 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/helpers/watcher.c
@@ -19,8 +19,8 @@
 #endif
 
 
-/** @defgroup watcher watcher: watch a specified variable and process
- *   it as an instance or scalar object
+/** @defgroup watcher watcher
+ *  Watch a specified variable and process it as an instance or scalar object
  *  @ingroup leaf
  *  @{
  */
@@ -217,7 +217,7 @@ netsnmp_watched_timestamp_handler(netsnmp_mib_handler *handler,
                                netsnmp_request_info *requests)
 {
     marker_t timestamp = (marker_t) handler->myvoid;
-    int      uptime;
+    long     uptime;
     int      cmp;
 
     DEBUGMSGTL(("helper:watcher:timestamp",
@@ -293,6 +293,7 @@ netsnmp_register_watched_spinlock(netsnmp_handler_registration *reginfo,
     netsnmp_inject_handler(reginfo, whandler);
     return netsnmp_register_watched_scalar(reginfo, winfo);
 }
+/**  @} */
 
 
 int
